requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
google-api-core==2.8.2
google-auth==2.8.0
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
numpy
pandas
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
rsa==4.8
scikit-learn
scipy
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
gunicorn==20.1.0
uvicorn